No Logic  Changes.  Renamed syntaxHighlight to syntaxHighlightOld
