+++ src/syntax-highlight/syntax-highlight-old.spec.js
